refactor(agent-commerce/A30): split [slug]/page.tsx into co-located helpers
