+++ src/app/agent-commerce/[slug]/page.tsx
@@ -20,8 +20,16 @@ import {
   getAgentCommerceItems,
   refreshAgentCommerceFromStore,
 } from "@/lib/agent-commerce";
-import type { AgentCommerceItem } from "@/lib/agent-commerce/types";
 import { getRepoProfile } from "@/lib/repo-profiles";
+
+import {
+  AisoScanSection,
+  LiveSignalsCard,
+  SCORE_LABELS,
+  findRelated,
+  getGradient,
+  getInitials,
+} from "./_sections";
 
 export const revalidate = 600;
 
@@ -42,51 +50,6 @@ export async function generateMetadata({
     alternates: { canonical: `/agent-commerce/${item.slug}` },
   };
 }
-
-function getInitials(name: string): string {
-  const parts = name.trim().split(/\s+/).filter(Boolean);
-  if (parts.length === 0) return "??";
-  if (parts.length === 1) return parts[0].slice(0, 2).toUpperCase();
-  return (parts[0][0] + parts[parts.length - 1][0]).toUpperCase();
-}
-
-function getGradient(name: string): string {
-  const grads = [
-    "linear-gradient(135deg, #3b82f6 0%, #1d4ed8 100%)",
-    "linear-gradient(135deg, #10b981 0%, #047857 100%)",
-    "linear-gradient(135deg, #f472b6 0%, #db2777 100%)",
-    "linear-gradient(135deg, #f59e0b 0%, #b45309 100%)",
-    "linear-gradient(135deg, #8b5cf6 0%, #6d28d9 100%)",
-    "linear-gradient(135deg, #06b6d4 0%, #0e7490 100%)",
-    "linear-gradient(135deg, #f97316 0%, #c2410c 100%)",
-  ];
-  let hash = 0;
-  for (const ch of name) hash = (hash * 31 + ch.charCodeAt(0)) >>> 0;
-  return grads[hash % grads.length];
-}
-
-function findRelated(
-  items: AgentCommerceItem[],
-  target: AgentCommerceItem,
-): AgentCommerceItem[] {
-  return items
-    .filter(
-      (it) =>
-        it.id !== target.id &&
-        (it.category === target.category || it.kind === target.kind),
-    )
-    .sort((a, b) => b.scores.composite - a.scores.composite)
-    .slice(0, 4);
-}
-
-const SCORE_LABELS: { key: keyof AgentCommerceItem["scores"]; label: string }[] = [
-  { key: "githubVelocity", label: "GitHub velocity" },
-  { key: "socialMentions", label: "Social mentions" },
-  { key: "pricingClarity", label: "Pricing clarity" },
-  { key: "apiClarity", label: "API clarity" },
-  { key: "aisoScore", label: "AISO score" },
-  { key: "portalReady", label: "Portal Ready" },
-];
 
 export default async function AgentCommerceDetailPage({ params }: DetailProps) {
   const { slug } = await params;
@@ -139,286 +102,9 @@ export default async function AgentCommerceDetailPage({ params }: DetailProps) {
         </div>
       </header>
 
-      {item.live ? (
-        <Card>
-          <CardHeader
-            showCorner
-            right={
-              item.live.fetchedAt ? (
-                <span>
-                  refreshed{" "}
-                  {(() => {
-                    const ms = Date.now() - new Date(item.live.fetchedAt).getTime();
-                    const mins = Math.floor(ms / 60_000);
-                    if (mins < 60) return `${mins}m ago`;
-                    const hrs = Math.floor(mins / 60);
-                    if (hrs < 48) return `${hrs}h ago`;
-                    return `${Math.floor(hrs / 24)}d ago`;
-                  })()}
-                </span>
-              ) : null
-            }
-          >
-            Live signals
-          </CardHeader>
-          <CardBody>
-            <div
-              style={{
-                display: "grid",
-                gridTemplateColumns: "repeat(auto-fit, minmax(140px, 1fr))",
-                gap: 14,
-                padding: "12px 14px",
-                fontFamily: "var(--font-mono, ui-monospace)",
-                fontSize: 12,
-              }}
-            >
-              {typeof item.live.stars === "number" ? (
-                <div>
-                  <div style={{ color: "var(--color-text-faint)", fontSize: 10, textTransform: "uppercase", letterSpacing: "0.06em" }}>
-                    Stars
-                  </div>
-                  <div style={{ color: "#fbbf24", fontSize: 20, fontWeight: 700 }}>
-                    {item.live.stars.toLocaleString("en-US")}
-                  </div>
-                </div>
-              ) : null}
-              {typeof item.live.forks === "number" ? (
-                <div>
-                  <div style={{ color: "var(--color-text-faint)", fontSize: 10, textTransform: "uppercase", letterSpacing: "0.06em" }}>
-                    Forks
-                  </div>
-                  <div style={{ color: "var(--color-text-default)", fontSize: 20, fontWeight: 700 }}>
-                    {item.live.forks.toLocaleString("en-US")}
-                  </div>
-                </div>
-              ) : null}
-              {item.live.pushedAt ? (
-                <div>
-                  <div style={{ color: "var(--color-text-faint)", fontSize: 10, textTransform: "uppercase", letterSpacing: "0.06em" }}>
-                    Last pushed
-                  </div>
-                  <div style={{ color: "#34d399", fontSize: 14, fontWeight: 700 }}>
-                    {(() => {
-                      const days = Math.max(
-                        0,
-                        Math.floor((Date.now() - new Date(item.live.pushedAt).getTime()) / 86_400_000),
-                      );
-                      return days === 0
-                        ? "today"
-                        : days === 1
-                          ? "1 day ago"
-                          : days < 30
-                            ? `${days} days ago`
-                            : days < 365
-                              ? `${Math.floor(days / 30)} months ago`
-                              : `${Math.floor(days / 365)} years ago`;
-                    })()}
-                  </div>
-                  <div style={{ color: "var(--color-text-faint)", fontSize: 10 }}>
-                    {new Date(item.live.pushedAt).toISOString().slice(0, 10)}
-                  </div>
-                </div>
-              ) : null}
-              {item.live.language ? (
-                <div>
-                  <div style={{ color: "var(--color-text-faint)", fontSize: 10, textTransform: "uppercase", letterSpacing: "0.06em" }}>
-                    Language
-                  </div>
-                  <div style={{ color: "var(--color-text-default)", fontSize: 14, fontWeight: 700 }}>
-                    {item.live.language}
-                  </div>
-                </div>
-              ) : null}
-              {typeof item.live.openIssues === "number" ? (
-                <div>
-                  <div style={{ color: "var(--color-text-faint)", fontSize: 10, textTransform: "uppercase", letterSpacing: "0.06em" }}>
-                    Open Issues
-                  </div>
-                  <div style={{ color: "var(--color-text-default)", fontSize: 14, fontWeight: 700 }}>
-                    {item.live.openIssues.toLocaleString("en-US")}
-                  </div>
-                </div>
-              ) : null}
-              {typeof item.live.hnMentions90d === "number" && item.live.hnMentions90d > 0 ? (
-                <div>
-                  <div style={{ color: "var(--color-text-faint)", fontSize: 10, textTransform: "uppercase", letterSpacing: "0.06em" }}>
-                    HN mentions (90d)
-                  </div>
-                  <div style={{ color: "#f97316", fontSize: 20, fontWeight: 700 }}>
-                    {item.live.hnMentions90d}
-                  </div>
-                  {item.live.hnTopUrl ? (
-                    <a
-                      href={item.live.hnTopUrl}
-                      target="_blank"
-                      rel="noreferrer"
-                      style={{ color: "var(--color-text-faint)", fontSize: 10, textDecoration: "none" }}
-                    >
-                      top story →
-                    </a>
-                  ) : null}
-                </div>
-              ) : null}
-              {typeof item.live.npmWeeklyDownloads === "number" && item.live.npmWeeklyDownloads > 0 ? (
-                <div>
-                  <div style={{ color: "var(--color-text-faint)", fontSize: 10, textTransform: "uppercase", letterSpacing: "0.06em" }}>
-                    npm /wk
-                  </div>
-                  <div style={{ color: "#cbd5e1", fontSize: 20, fontWeight: 700 }}>
-                    {item.live.npmWeeklyDownloads >= 1_000_000
-                      ? `${(item.live.npmWeeklyDownloads / 1_000_000).toFixed(2)}M`
-                      : item.live.npmWeeklyDownloads >= 1000
-                        ? `${(item.live.npmWeeklyDownloads / 1000).toFixed(1)}k`
-                        : `${item.live.npmWeeklyDownloads}`}
-                  </div>
-                  {item.live.npmName ? (
-                    <a
-                      href={item.live.npmRegistryUrl}
-                      target="_blank"
-                      rel="noreferrer"
-                      style={{ color: "var(--color-text-faint)", fontSize: 10, textDecoration: "none" }}
-                    >
-                      {item.live.npmName} {item.live.npmLatestVersion ? `@${item.live.npmLatestVersion}` : ""} →
-                    </a>
-                  ) : null}
-                </div>
-              ) : null}
-            </div>
-          </CardBody>
-        </Card>
-      ) : null}
+      {item.live ? <LiveSignalsCard live={item.live} /> : null}
 
-      {aisoScan && aisoScan.status === "completed" ? (
-        <Card>
-          <CardHeader
-            showCorner
-            right={
-              aisoScan.tier ? (
-                <span style={{ textTransform: "uppercase", letterSpacing: "0.06em" }}>
-                  {aisoScan.tier}
-                </span>
-              ) : null
-            }
-          >
-            AS · AISO score
-          </CardHeader>
-          <CardBody>
-            <div style={{ padding: "12px 14px", display: "grid", gap: 14 }}>
-              <div style={{ display: "flex", alignItems: "baseline", gap: 12 }}>
-                <span
-                  style={{
-                    fontFamily: "var(--font-mono, ui-monospace)",
-                    fontSize: 32,
-                    fontWeight: 700,
-                    color: "var(--color-text-default)",
-                    lineHeight: 1,
-                  }}
-                >
-                  {aisoScan.score ?? "—"}
-                </span>
-                {aisoScan.tier ? (
-                  <span
-                    style={{
-                      fontFamily: "var(--font-mono, ui-monospace)",
-                      fontSize: 12,
-                      color: "var(--color-text-faint)",
-                      textTransform: "uppercase",
-                      letterSpacing: "0.06em",
-                    }}
-                  >
-                    · {aisoScan.tier}-tier
-                  </span>
-                ) : null}
-              </div>
-
-              {aisoScan.dimensions.length > 0 ? (
-                <div className="ac-score-rows">
-                  {aisoScan.dimensions.map((d) => {
-                    const n = Math.max(0, Math.min(100, d.score));
-                    return (
-                      <div className="ac-score-row" key={d.key}>
-                        <span>{d.label}</span>
-                        <span className="ac-score-track">
-                          <i style={{ width: `${n}%` }} />
-                        </span>
-                        <span className="ac-score-num">{n}</span>
-                      </div>
-                    );
-                  })}
-                </div>
-              ) : null}
-
-              {aisoScan.issues.length > 0 ? (
-                <ul
-                  style={{
-                    margin: 0,
-                    padding: 0,
-                    listStyle: "none",
-                    display: "grid",
-                    gap: 4,
-                    fontFamily: "var(--font-mono, ui-monospace)",
-                    fontSize: 11,
-                  }}
-                >
-                  {aisoScan.issues.slice(0, 5).map((issue, i) => (
-                    <li
-                      key={`${issue.severity}-${i}`}
-                      style={{ color: "var(--color-text-default)" }}
-                    >
-                      <span
-                        style={{
-                          color: "var(--color-text-faint)",
-                          textTransform: "uppercase",
-                          letterSpacing: "0.06em",
-                          marginRight: 8,
-                        }}
-                      >
-                        [{issue.severity}]
-                      </span>
-                      {issue.title}
-                    </li>
-                  ))}
-                </ul>
-              ) : null}
-
-              <div
-                style={{
-                  fontFamily: "var(--font-mono, ui-monospace)",
-                  fontSize: 10,
-                  color: "var(--color-text-faint)",
-                  textTransform: "uppercase",
-                  letterSpacing: "0.06em",
-                }}
-              >
-                Last scanned{" "}
-                {aisoScan.completedAt
-                  ? new Date(aisoScan.completedAt).toISOString().slice(0, 10)
-                  : "—"}
-              </div>
-            </div>
-          </CardBody>
-        </Card>
-      ) : aisoScan && (aisoScan.status === "queued" || aisoScan.status === "running") ? (
-        <div
-          style={{
-            display: "inline-flex",
-            alignSelf: "flex-start",
-            alignItems: "center",
-            gap: 6,
-            padding: "4px 10px",
-            borderRadius: 999,
-            border: "1px solid var(--color-border, rgba(255,255,255,0.12))",
-            background: "var(--color-bg-soft, rgba(255,255,255,0.03))",
-            fontFamily: "var(--font-mono, ui-monospace)",
-            fontSize: 10,
-            textTransform: "uppercase",
-            letterSpacing: "0.08em",
-            color: "var(--color-text-faint)",
-          }}
-        >
-          AISO scan in progress
-        </div>
-      ) : null}
+      {aisoScan ? <AisoScanSection aisoScan={aisoScan} /> : null}
 
       <div className="ac-detail-grid">
         <div style={{ display: "grid", gap: 12 }}>
